+++ test/cli-build/providers/tsts-type-forms/core.test.mjs
@@ -357,7 +357,7 @@ test("CLI consumes TSTS as-const literal readonly results and rejects readonly w
   ]);
 
   assert.match(generatedSource, /public static string status\(\)/);
-  assert.match(generatedSource, /\(string, double\) row = \("ready", 7\);/);
+  assert.match(generatedSource, /\(string, int\) row = \("ready", 7\);/);
   assert.match(generatedSource, /return \$"\{row\.Item1\}:\{row\.Item2\}";/);
   assert.doesNotMatch(generatedSource, /as const/);
   assert.doesNotMatch(generatedSource, /__unsupported/);
